fix(csa-executor): harden ACP crash exit-code extraction
